EPMRPP-96415 || Default number of items per page is displayed when user returns to the page
